progress
I have filled in the User class and the WorkOrder class. I think there are some additional values that we wanted for the work order, but I wasn't exactly clear on what those were. I have also done the login xml and activity class.  The login screen is very basic. User name and password with a button to login and an option at the bottom to create a new user. Please feel free to make any additional changes. I will try working on the create user screen and activity later.